requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# api/__init__.py: 12,13,14
# api/common/base.py: 11,12
# api/common/decorator.py: 8,9,10,11
# api/service.py: 6,7
Flask == 0.12

# api/__init__.py: 16
Flask_Cors == 3.0.2

# api/__init__.py: 15
Werkzeug == 0.11.13

# api/__init__.py: 17
# api/common/base.py: 14
# api/common/decorator.py: 15
# api/local_config.py: 7
# api/run.py: 8,10,11,12
# api/service.py: 8,9,10,11,12
# run_api.py: 8
api.ai == 0.1.0

# api/__init__.py: 17
# api/common/base.py: 14
# api/common/decorator.py: 15
# api/local_config.py: 7
# api/run.py: 8,10,11,12
# api/service.py: 8,9,10,11,12
# run_api.py: 8
bgionline == 2.1.3

# api/__init__.py: 17
# api/common/base.py: 14
# api/common/decorator.py: 15
# api/local_config.py: 7
# api/run.py: 8,10,11,12
# api/service.py: 8,9,10,11,12
# run_api.py: 8
bigjob == 0.64.5

# api/__init__.py: 17
# api/common/base.py: 14
# api/common/decorator.py: 15
# api/local_config.py: 7
# api/run.py: 8,10,11,12
# api/service.py: 8,9,10,11,12
# run_api.py: 8
browserstack == 0.1

# api/__init__.py: 17
# api/common/base.py: 14
# api/common/decorator.py: 15
# api/local_config.py: 7
# api/run.py: 8,10,11,12
# api/service.py: 8,9,10,11,12
# run_api.py: 8
epguides-api == 1.0.1

# api/__init__.py: 17
# api/common/base.py: 14
# api/common/decorator.py: 15
# api/local_config.py: 7
# api/run.py: 8,10,11,12
# api/service.py: 8,9,10,11,12
# run_api.py: 8
nvpnsxapi == 0.1.3.dev1

# api/__init__.py: 17
# api/common/base.py: 14
# api/common/decorator.py: 15
# api/local_config.py: 7
# api/run.py: 8,10,11,12
# api/service.py: 8,9,10,11,12
# run_api.py: 8
openstackocci == essex-1.1

# api/__init__.py: 11
pyahocorasick == 1.1.4

# api/__init__.py: 17
# api/common/base.py: 14
# api/common/decorator.py: 15
# api/local_config.py: 7
# api/run.py: 8,10,11,12
# api/service.py: 8,9,10,11,12
# run_api.py: 8
saga-bigjob == 0.50b-10-gfdf8243

# api/__init__.py: 17
# api/common/base.py: 14
# api/common/decorator.py: 15
# api/local_config.py: 7
# api/run.py: 8,10,11,12
# api/service.py: 8,9,10,11,12
# run_api.py: 8
ucloud-client == 1.0
